test: split billing cycle state tests
Changed:
Split the billing cycle state transition tests into focused files and moved shared test setup into a local helper module.

Benefit:
Billing cycle state coverage is easier to review and maintain without changing production behavior.
